requirements.txt
cffi==1.3.0
cryptography==1.0.2
enum34==1.0.4
Flask==0.10.1
Flask-MySQL==1.3
idna==2.0
ipaddress==1.0.14
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mysql==0.0.1
MySQL-python==1.2.5
mysqlclient==1.3.6
pyasn1==0.1.9
pycparser==2.14
six==1.10.0
Werkzeug==0.10.4
wheel==0.24.0
